==> 03. Adding_private_sequence_data.Rmd <==
---
title: "Creating customised DNA reference databases - private sequence data"
author: "Jessica Chung /Melissa Carew"
date: "15/02/2024"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

--------------------------------------------------------------------------------
## SECTION 3: Preparing a private library as a DNA reference database

This step imports DNA barcodes which I have produced from invertebrates. It is a mixture of unpublished and published data. The private reference library is managed in Geneious prime and can be exported out as .csv. Private libraries can be imported into this documant as .csv files and need to contain at a minimum: The lowest possible taxonomic classification (superkingdom;phylum;class;order;family;genus;species), an individual sequence identifier (i.e., sequence, sample or voucher code) and a COI DNA sequence. Other metadata can be included but will not be used for reference library construction.

Please refer to the private library checking scripts (in 'A1. Private_database_checks_corrections.Rmd' workflow document) before adding private sequence data. This document can be used to check data for errors.

# read in DNA barcode datafile in .csv 
```{r}
priv_lib <- read_csv (here::here("raw_libraries/priv_lib_Nov23.csv"))
```

# set fields for refdb
```{r}
priv_lib <- refdb::refdb_set_fields(priv_lib,
                             taxonomy = c(kingdom ="superkingdom",
                               phylum = "phylum",
                               class = "class",
                               order = "order",
                               family = "family",
                               genus = "genus",
                               species = "species"),
                             sequence = "DNA_seq",
                             marker = "marker",
                             source = "source",
                             id = "id")

refdb::refdb_get_fields(priv_lib)
```
#Step 1: Remove duplicates.
```{r}
priv_lib_clean <- refdb::refdb_filter_seq_duplicates(priv_lib)

rm(priv_lib)
```

# plot sequence length
```{r}
refdb::refdb_plot_seqlen_hist(priv_lib_clean)
```

Note: Primer cropping in refdb does not work! I could not get refdb to properly trim primers it needed really high error setting to do any trimming, but with these settings over trimmed most of the data. for the short amplicon the forward primer trims ok but the reverse trims in the wrong spot.

#Visualizing the reference library (Optional)

Now let’s take a tour of the functions you can use to produce graphical representation of your reference database. Because refdb stores reference database as dataframes it is straightforward to produce plots (e.g. with the tidyverse and ggplot2). For example, we can make a barplot showing the distribution of the families like this:
  
```{r}
priv_lib_clean %>% 
  group_by(phylum) %>% 
  count() %>% 
  ggplot2::ggplot(aes(fct_reorder(phylum, n, .desc = TRUE), n)) +
  geom_col() +
  xlab("Phylum") +
  ylab("Number of records") +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
```

```{r}
priv_lib_clean %>% 
  group_by(class) %>% 
  count() %>% 
  ggplot(aes(fct_reorder(class, n, .desc = TRUE), n)) +
  geom_col() +
  xlab("Class") +
  ylab("Number of records") +
  theme(axis.text.x = element_text(angle = 90, vjust = 1, hjust=1))
```

Additionally, refdb provides ready-to-use functions to produce more sophisticated plots. For example, to represent the taxonomic coverage of your reference library across multiple levels with a tree map:
  
Jess - This does not work but not important as there is simply too much data to get anything meaningful out.

```{r}
refdb::refdb_plot_tax_treemap(priv_lib_clean, cols = 4)
```

Alternatively you can represent this information with a taxonomic tree. These functions have several parameters to control what is represented in the plot (taxonomic ranks, colors, etc.).

```{r}
refdb::refdb_plot_tax_tree(priv_lib_clean)
```

# Creating a report
A report is a simple and rapid way to get an overview of the current state of your reference library and to identify some possible issues. You can compile a report using the function refdb_report: 
  
```{r}
refdb::refdb::refdb_report(priv_lib_clean, view=TRUE)
```                    

# Exporting cleaned private data
Once satisfied with the results, libraries can be exported for further use. Since the reference database is a dataframe, it is possible to use any export functions (from R or from other packages) that support dataframes.

```{r}
# importing library into R (if required)
saveRDS(priv_lib_clean,(here::here("cleaned_libraries/priv_lib_clean_Oct23.rds")))

# reload all private document for cleaning (if needed)
#priv_lib_clean <- readRDS(here::here("cleaned_libraries/priv_lib_clean_Oct23.rds"))

```

# set fields for priv_lib_clean
```{r}
priv_lib_clean <- refdb::refdb_set_fields(priv_lib_clean,
                             taxonomy = c(kingdom = "superkingdom",
                               phylum = "phylum",
                               class = "class",
                               order = "order",
                               family = "family",
                               genus = "genus",
                               species = "species"),
                             sequence = "DNA_seq",
                             marker = "marker",
                             source = "source",
                             id = "id")

refdb::refdb_get_fields(priv_lib_clean)
```
